+++ src/plugins/PNP_BrightWave.h
@@ -48,7 +48,7 @@ public:
     angleNext = PI_VALUE; // starting angle for minimal brightness
   }
 
-  void trigger(PixelNutHandle handle, PixelNutSupport::DrawProps *pdraw, short force)
+  void trigger(PixelNutHandle handle, PixelNutSupport::DrawProps *pdraw, byte force)
   {
     forceVal = force;
   }
@@ -80,7 +80,7 @@ public:
 
 private:
   uint16_t myid;
-  short forceVal;
+  byte forceVal;
   uint16_t baseValue;
   float angleNext;
 };
