+++ src/plugins/PNP_DelayWave.h
@@ -47,7 +47,7 @@ public:
     angleNext = 0.0;  // starting angle
   }
 
-  void trigger(PixelNutHandle handle, PixelNutSupport::DrawProps *pdraw, short force)
+  void trigger(PixelNutHandle handle, PixelNutSupport::DrawProps *pdraw, byte force)
   {
     forceVal = force;
   }
@@ -78,7 +78,7 @@ public:
 
 private:
   uint16_t myid;
-  short forceVal;
+  byte forceVal;
   uint16_t maxDelay;
   float angleNext;
 };
