+++ src/xplugins/PNP_Spectra.h
@@ -133,7 +133,7 @@ public:
     }
   }
 
-  void trigger(PixelNutHandle handle, PixelNutSupport::DrawProps *pdraw, short force)
+  void trigger(PixelNutHandle handle, PixelNutSupport::DrawProps *pdraw, byte force)
   {
     if (hueVals != NULL)
     {
